Move UrlValidation out into its own class
