+++ lib/html/proofer/utils.rb
@@ -5,7 +5,12 @@ module HTML
     extend self
 
     def create_nokogiri(path)
-      content = File.open(path).read
+      if File.exist? path
+        content = File.open(path).read
+      else
+        content = path
+      end
+
       Nokogiri::HTML(content)
     end
   end
